refactor(loader): extract fast-path logic into ep-gen-code-bundle.ts
Move LOADER_CACHE_BUST, LOADER_ASSETS_BUCKET, makeCodegenBucketPath,
makeBundleBucketPath, makeExportOptsKey, and tryGetCachedPublishedBundle
into ep-gen-code-bundle.ts. gen-code-bundle.ts now calls the function
directly and re-exports the public constants for backward compat.
